requirements.txt
pandas
Django
requests
djangorestframework